class - Prop Drilling no React
