+++ src/pages/Home/components/Countdown/styles.ts
