+++ src/pages/Home/components/NewCycleForm/styles.ts
